requirements_checkpoint.txt
contourpy==1.3.2
cycler==0.12.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.9.0
Jinja2==3.1.6
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
ml_dtypes==0.5.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
onnx==1.19.1
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
platformdirs==4.5.0
protobuf==6.32.1
pycuda==2025.1.2
pyparsing==3.2.5
pyrealsense2==2.56.5.9235
pyserial==3.5
python-dateutil==2.9.0.post0
pytools==2025.2.5
siphash24==1.8
six==1.17.0
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
typing_extensions==4.15.0
ydlidar==1.2.4
